Time: 5 ms (33%), Space: 69.8 MB (25.6%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,6 +1,6 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
-        double n= nums.length;
+        int n= nums.length;
         double sum =0;
         double  max= Integer.MIN_VALUE;
